Resolve "BIND 9.16 unit tests failing reliably on x86_64 NUMA machines"

Closes #2396

See merge request isc-projects/bind9!4687

+++ lib/isc/win32/include/isc/condition.h
@@ -22,8 +22,8 @@
 typedef struct isc_condition_thread isc_condition_thread_t;
 
 struct isc_condition_thread {
-	unsigned long th;
-	HANDLE	      handle[2];
+	uintptr_t th;
+	HANDLE	  handle[2];
 	ISC_LINK(isc_condition_thread_t) link;
 };
